refactor: split out os-specific code
